edge/detector2/requirements.txt
absl-py
astunparse
cachetools
certifi
chardet
cycler
ffmpeg-python
future
gast
google-auth
google-auth-oauthlib
google-pasta
googledrivedownloader
grpcio
h5py
idna
importlib-metadata
imutils
Keras-Preprocessing
kiwisolver
Markdown
matplotlib
numpy
oauthlib
#opencv-contrib-python
#opencv-python
opt-einsum
pandas
Pillow
protobuf
pyasn1
pyasn1-modules
pyparsing
PyQt5
PyQt5-sip
python-dateutil
pytz
PyYAML
requests
requests-oauthlib
rsa
scipy
six
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
tqdm
urllib3
Werkzeug
wrapt
zipp
